refactoring: created utils module; added mask_clouds draft
